Music stuff

+++ src/main/java/net/kodehawa/mantarobot/commands/music/TrackScheduler.java
@@ -92,12 +92,12 @@ public class TrackScheduler extends AudioEventAdapter {
     }
     private void announce() {
         try {
-            if (getPreviousTrack().getRequestedChannel() != null && getPreviousTrack().getRequestedChannel().canTalk())
+            if (getPreviousTrack() != null && getPreviousTrack().getRequestedChannel() != null && getPreviousTrack().getRequestedChannel().canTalk())
                 getPreviousTrack().getRequestedChannel().getMessageById(lastAnnounce).complete().delete().queue();
         } catch (Exception ignored) {
         }
         try {
-            if (getCurrentTrack().getRequestedChannel() != null && getCurrentTrack().getRequestedChannel().canTalk())
+            if (getCurrentTrack() != null && getCurrentTrack().getRequestedChannel() != null && getCurrentTrack().getRequestedChannel().canTalk())
                 getCurrentTrack().getRequestedChannel()
                         .sendMessage("\uD83D\uDCE3 Now playing in " + getAudioManager().getConnectedChannel().getName()
                                 + ": " + getCurrentTrack().getInfo().title + " (" + AudioUtils.getLength(getCurrentTrack().getInfo().length) + ")"
@@ -121,6 +121,8 @@ public class TrackScheduler extends AudioEventAdapter {
             if (repeat == Repeat.QUEUE)
                 queue.offer(previousTrack);
         }
+        if (currentTrack == null)
+            onTrackSchedulerStop();
     }
 
     public void shuffle() {
@@ -153,5 +155,11 @@ public class TrackScheduler extends AudioEventAdapter {
         if (getCurrentTrack().getRequestedChannel() != null && getCurrentTrack().getRequestedChannel().canTalk())
             getCurrentTrack().getRequestedChannel().sendMessage("Track got stuck, skipping...").queue();
         next(true);
+    }
+
+    private void onTrackSchedulerStop() {
+        getGuild().getAudioManager().closeAudioConnection();
+        if (getPreviousTrack().getRequestedChannel() != null && getPreviousTrack().getRequestedChannel().canTalk())
+            getPreviousTrack().getRequestedChannel().sendMessage(":mega: Finished playing queue.").queue();
     }
 }
